12-projects/level-200/climate/airflow/requirements.txt
apache-airflow-providers-google
apache-airflow-providers-postgres
pyarrow
pyspark==3.2.0
pandas
